requirements.txt
langchain==0.1.16
streamlit==1.33.0
langchain-openai==0.1.3
openai==1.17.1
langchain-community==0.0.32
langchain-text-splitters==0.0.1
python-dotenv==1.0.1
boto3==1.34.84
langchain-core==0.1.42
faiss-cpu==1.8.0
python-doctr==0.8.1
tf2onnx==1.16.1
tensorflow==2.15.0
